+++ controllers/SubDivision.js
@@ -383,7 +383,7 @@ const getSubdivisionAreaPercentages = async (_req, res) => {
             subdiv_code,
             subdiv_name,
             ROUND(
-                SUM(district_area) / (SELECT SUM(district_area) FROM normal_district_details) * 100,
+                (SUM(district_area) / (SELECT SUM(district_area) FROM normal_district_details) * 100)::numeric,
                 2
             ) AS area_percentage
         FROM normal_district_details
